feat: create hcc db services
