+++ src/db/tests/legallead.jdbc.tests/implementations/HarrisCriminalZipFileReaderTests.cs
@@ -1,5 +1,4 @@
-﻿using Castle.Core.Configuration;
-using legallead.jdbc.helpers;
+﻿using legallead.jdbc.helpers;
 using legallead.jdbc.implementations;
 using legallead.jdbc.interfaces;
 using Microsoft.Extensions.DependencyInjection;
@@ -8,12 +7,25 @@ using System.Diagnostics;
 
 namespace legallead.jdbc.tests.implementations
 {
-    public class HarrisCriminalYearToDateReaderTests
+    public class HarrisCriminalZipFileReaderTests
     {
+        [Theory]
+        [InlineData("")]
+        [InlineData("not-an-existing-file-path")]
+        public void ServiceReadRequiresFileName(string sourceFile)
+        {
+            Assert.Throws<InvalidOperationException>(() =>
+            {
+                using var service = GetReader(sourceFile);
+                service.Read();
+            });
+        }
+
         [Theory]
         [InlineData(historyData)]
         public void ServiceCanReadZipContent(string sourceFile)
         {
+            if (!Debugger.IsAttached) return;
             var error = Record.Exception(() =>
             {
                 using var service = GetReader(sourceFile);
@@ -25,6 +37,7 @@ namespace legallead.jdbc.tests.implementations
         [InlineData(historyData)]
         public void ServiceCanTranslateZipContent(string sourceFile)
         {
+            if (!Debugger.IsAttached) return;
             var error = Record.Exception(() =>
             {
                 using var service = GetReader(sourceFile);
@@ -37,6 +50,7 @@ namespace legallead.jdbc.tests.implementations
         [InlineData(historyData)]
         public void ServiceCanTransferZipContent(string sourceFile)
         {
+            if (!Debugger.IsAttached) return;
             var error = Record.Exception(() =>
             {
                 using var service = GetReader(sourceFile);
@@ -54,13 +68,13 @@ namespace legallead.jdbc.tests.implementations
             });
             Assert.Null(error);
         }
-        private static HarrisCriminalFileReader GetReader(string sourceFile)
+        private static HarrisCriminalZipFileReader GetReader(string sourceFile)
         {
             var response = new KeyValuePair<bool, string>(true, "test");
             var completion = Task.FromResult(response);
             var mock = new Mock<IHarrisLoadRepository>();
             mock.Setup(x => x.Append(It.IsAny<string>())).Returns(completion);
-            return new HarrisCriminalFileReader(sourceFile, mock.Object);
+            return new HarrisCriminalZipFileReader(sourceFile, mock.Object);
         }
 
 
@@ -72,7 +86,7 @@ namespace legallead.jdbc.tests.implementations
             if (!Debugger.IsAttached) return;
             var error = Record.Exception(() =>
             {
-                using var service = new HarrisCriminalFileReader(sourceFile, svc);
+                using var service = new HarrisCriminalZipFileReader(sourceFile, svc);
                 service.Transfer();
             });
             Assert.Null(error);
